Start on GUI
I got the basic setup for the home page and it redirects to the other
page. Also the "No" button redirects back to the main page, but I'm not
sure why. My next step is to try to make the picture show up in the app
because it won't when I run it on my phone.